Use proper alignment and padding for pointer arguments -- runtime.
This CL modifies the shared memory allocation for pointer arguments to
use proper alignment and padding. We use default alignment (64) and
min padding (1) for HIDL drivers, and default alignment (64) and default
padding (64) for sAIDL drivers.

Bug: 184164929
Test: NNT_static

+++ common/TypeUtils.cpp
@@ -212,7 +212,7 @@ std::pair<size_t, std::vector<size_t>> getMemorySizes(const Model& model) {
     return std::make_pair(operandValuesSize, std::move(poolSizes));
 }
 
-uint32_t roundUp(uint32_t size, uint32_t multiple) {
+size_t roundUp(size_t size, size_t multiple) {
     CHECK(multiple != 0);
     CHECK((multiple & (multiple - 1)) == 0) << multiple << " is not a power of two";
     return (size + (multiple - 1)) & ~(multiple - 1);
